A class for converting a dev file to a netcdf grouping.
This class creates a netcdf that has a group called 'factory_calibration' that contains calibration info output by the pyACS.ACS class. For ACS datasets that are saved as netCDFs, this will allow the dev file calibration information to accompany the data as a separate grouping within the same file.

See the accompanying example for what the output would look like. I would appreciate discussion about the appropriate vocabulary.

I'm amenable to turning this into a flat function rather than a class.