Added class for ingesting .dev files.
At class instantiation, file metadata is assigned as class attributes. With some care, this could be used to replace the read_device_file and set_frame_descriptor functions in the base ACS class.

Additionally, functions are supplied that export metadata to appropriately shaped arrays  in xarray datasets and netcdfs. These are useful for performing advanced operations on ACS data. A brief example is provided in the examples folder.